improve perf of rendering error log items
